Add dummy UTP callbacks.
Libutp will sometimes call our callbacks after we called UTP_Close,
notably to notify us of the UTP_STATE_DESTROYING state change, but
also, for some reason, to ask us about our read buffer.  The simplest
way to avoid issues with that is to switch to a second set of callbacks.

git-svn-id: svn://svn.transmissionbt.com/Transmission@11910 f4695dd4-2c0a-0410-b89c-da849a56a58e