(trunk libT) better scrape management on private trackers.
If a private tracker scrape says that there are no downloaders in the swarm, mark all the peers in the private swarm as seeds. This can greatly reduce unnecessary overhead on large seedboxes. We don't do this same trick on public torrents, since a public tracker won't know of all the peers.


git-svn-id: svn://svn.transmissionbt.com/Transmission@12253 f4695dd4-2c0a-0410-b89c-da849a56a58e