(trunk libT) don't cache unininteresting peer addresses between sessions
For example, if we're both seeds, or if the peer is not connectible, don't bother caching it for the next session. If it's still alive, we'll find it up through DHT or tracker announces next time around. As with r12253, this commit's intention is to reduce the number of unproductive peer_atoms that we waste time trying to connect to.


git-svn-id: svn://svn.transmissionbt.com/Transmission@12254 f4695dd4-2c0a-0410-b89c-da849a56a58e