Silence a minor compiler warning.
We used to pass no button text + response_id pairs when creating the dialog because only gtk_dialog_add_button() would return back the actual button widget s.t. we could use a grab on it. However as of GTK+ 2.20 we can use gtk_dialog_get_widget_for_response() to get the button, so we can create dialogs in the more common way (which also fixes the gcc warning)


git-svn-id: svn://svn.transmissionbt.com/Transmission@13157 f4695dd4-2c0a-0410-b89c-da849a56a58e