Improve the look of torrents with errors a bit (Qt client)
Display emblem over mime icon if torrent has error(s) to indicate the
issue.
Use normal text color instead of red when drawing selected item: there
is no easy way to be sure that red color looks nice on selection
background, and most of the time it doesn't. Using red color for non-
selected items is questionable too, but let's leave it at that for now.


git-svn-id: svn://svn.transmissionbt.com/Transmission@14452 f4695dd4-2c0a-0410-b89c-da849a56a58e